fix cbrt on win32
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@47523 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: a34fd0959ce6b58e0d7e9f0df2326fe95d9cc5e0

+++ LHCb/Kernel/LHCbMath/LHCbMath/Power.h
@@ -92,7 +92,7 @@ namespace Gaudi
 #ifdef __GNUC__
       return ::cbrtf ( __x           );
 #else
-      return ::pow   ( __x , (double)1.0/3.0 );
+      return ::pow   ( __x , (float)1.0/3.0 );
 #endif
     }  
 
@@ -113,7 +113,7 @@ namespace Gaudi
 #ifdef __GNUC__
       return ::cbrtl ( __x           );
 #else
-      return ::pow   ( __x , (double)1.0/3.0 );
+      return ::pow   ( __x , (long double)1.0/3.0 );
 #endif
     }
